update: new separate package for account and session
