+++ internal/container/usecase.go
@@ -2,17 +2,20 @@ package container
 
 import (
 	"github.com/DucTran999/auth-service/internal/handler/background"
-	"github.com/DucTran999/auth-service/internal/usecase"
+	"github.com/DucTran999/auth-service/internal/usecase/account"
 	"github.com/DucTran999/auth-service/internal/usecase/auth"
 	"github.com/DucTran999/auth-service/internal/usecase/port"
+	"github.com/DucTran999/auth-service/internal/usecase/session"
 	"github.com/DucTran999/auth-service/internal/usecase/shared"
 )
 
 type useCases struct {
-	jwtAuth           port.JWTAuthUsecase
-	sessionAuth       port.AuthUseCase
-	account           port.AccountUseCase
-	restSession       port.SessionUsecase
+	jwtAuth     port.JWTAuthUsecase
+	sessionAuth port.SessionAuthUseCase
+
+	account     port.AccountUseCase
+	restSession port.SessionUsecase
+
 	backgroundSession background.SessionUsecase
 }
 
@@ -21,7 +24,7 @@ func (c *Container) initUseCases() {
 		c.Hasher,
 		c.repositories.account,
 	)
-	accountUC := usecase.NewAccountUseCase(
+	accountUC := account.NewAccountUseCase(
 		c.Hasher,
 		c.repositories.account,
 	)
@@ -35,7 +38,7 @@ func (c *Container) initUseCases() {
 	)
 	jwtAuthUC := auth.NewAuthJWTUsecase(c.Signer, c.Cache, accountVerifier)
 
-	sessionUC := usecase.NewSessionUC(c.Cache, c.repositories.session)
+	sessionUC := session.NewSessionUC(c.Cache, c.repositories.session)
 
 	c.useCases = &useCases{
 		account:           accountUC,
